// backend/src/generated/chat_pb.d.ts
export class ConnectRequest extends jspb.Message {
  getUserId(): number
  setUserId(value: number): void

  getActive(): boolean
  setActive(value: boolean): void

  serializeBinary(): Uint8Array
  toObject(includeInstance?: boolean): ConnectRequest.AsObject
  static toObject(
    includeInstance: boolean,
    msg: ConnectRequest,
  ): ConnectRequest.AsObject
  static extensions: { [key: number]: jspb.ExtensionFieldInfo<jspb.Message> }
  static extensionsBinary: {
    [key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>
  }
  static serializeBinaryToWriter(
    message: ConnectRequest,
    writer: jspb.BinaryWriter,
  ): void
  static deserializeBinary(bytes: Uint8Array): ConnectRequest
  static deserializeBinaryFromReader(
    message: ConnectRequest,
    reader: jspb.BinaryReader,
  ): ConnectRequest
}

export namespace ConnectRequest {
  export type AsObject = {
    userId: number
    active: boolean
  }
}
// ...
export class Message extends jspb.Message {
  getId(): number
  setId(value: number): void

  getContent(): string
  setContent(value: string): void

  getTimestamp(): string
  setTimestamp(value: string): void

  getUserId(): number
  setUserId(value: number): void

  serializeBinary(): Uint8Array
  toObject(includeInstance?: boolean): Message.AsObject
  static toObject(includeInstance: boolean, msg: Message): Message.AsObject
  static extensions: { [key: number]: jspb.ExtensionFieldInfo<jspb.Message> }
  static extensionsBinary: {
    [key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>
  }
  static serializeBinaryToWriter(
    message: Message,
    writer: jspb.BinaryWriter,
  ): void
  static deserializeBinary(bytes: Uint8Array): Message
  static deserializeBinaryFromReader(
    message: Message,
    reader: jspb.BinaryReader,
  ): Message
}

export namespace Message {
  export type AsObject = {
    id: number
    content: string
    timestamp: string
    userId: number
  }
}